@@ -1,9 +1,8 @@
 module github.com/teamwork/reload
 
-go 1.12
+go 1.13
 
 require (
 	github.com/fsnotify/fsnotify v1.4.7
-	github.com/pkg/errors v0.8.1
 	golang.org/x/sys v0.0.0-20191110163157-d32e6e3b99c4 // indirect
 )
